% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmapGridDataPlot.R, R/tmapLeafletDataPlot.R,
%   R/tmapMapboxDataPlot.R, R/tmapMaplibreDataPlot.R
\name{tmapGridDataPlot.tm_data_cartogram}
\alias{tmapGridDataPlot.tm_data_cartogram}
\alias{tmapGridDataPlot.tm_data_cartogram_ncont}
\alias{tmapGridDataPlot.tm_data_cartogram_dorling}
\alias{tmapLeafletDataPlot.tm_data_cartogram}
\alias{tmapLeafletDataPlot.tm_data_cartogram_ncont}
\alias{tmapLeafletDataPlot.tm_data_cartogram_dorling}
\alias{tmapMapboxDataPlot.tm_data_cartogram}
\alias{tmapMapboxDataPlot.tm_data_cartogram_ncont}
\alias{tmapMapboxDataPlot.tm_data_cartogram_dorling}
\alias{tmapMaplibreDataPlot.tm_data_cartogram}
\alias{tmapMaplibreDataPlot.tm_data_cartogram_ncont}
\alias{tmapMaplibreDataPlot.tm_data_cartogram_dorling}
\title{Intermal methods}
\usage{
\method{tmapGridDataPlot}{tm_data_cartogram}(
  a,
  shpTM,
  dt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapGridDataPlot}{tm_data_cartogram_ncont}(
  a,
  shpTM,
  dt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapGridDataPlot}{tm_data_cartogram_dorling}(
  a,
  shpTM,
  dt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapLeafletDataPlot}{tm_data_cartogram}(
  a,
  shpTM,
  dt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapLeafletDataPlot}{tm_data_cartogram_ncont}(
  a,
  shpTM,
  dt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapLeafletDataPlot}{tm_data_cartogram_dorling}(
  a,
  shpTM,
  dt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

tmapMapboxDataPlot.tm_data_cartogram(
  a,
  shpTM,
  dt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

tmapMapboxDataPlot.tm_data_cartogram_ncont(
  a,
  shpTM,
  dt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

tmapMapboxDataPlot.tm_data_cartogram_dorling(
  a,
  shpTM,
  dt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

tmapMaplibreDataPlot.tm_data_cartogram(
  a,
  shpTM,
  dt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

tmapMaplibreDataPlot.tm_data_cartogram_ncont(
  a,
  shpTM,
  dt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

tmapMaplibreDataPlot.tm_data_cartogram_dorling(
  a,
  shpTM,
  dt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)
}
\description{
Intermal methods
}
\keyword{internal}
