% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmapTransCartogram.R
\name{tmapTransCartogram}
\alias{tmapTransCartogram}
\title{Cartogram transformation}
\usage{
tmapTransCartogram(shpTM, size, ord__, plot.order, args, scale)
}
\arguments{
\item{shpTM, size, ord__, plot.order, args, scale}{tmap internals}
}
\description{
Cartogram transformation
}
\keyword{internal}
