% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_layers_cartogram.R
\name{tm_cartogram}
\alias{tm_cartogram}
\alias{tm_cartogram_ncont}
\alias{tm_cartogram_dorling}
\alias{opt_tm_cartogram}
\alias{opt_tm_cartogram_ncont}
\alias{opt_tm_cartogram_dorling}
\title{Map layer: cartogram}
\usage{
tm_cartogram(
  size = 1,
  size.scale = tmap::tm_scale(),
  size.legend = tmap::tm_legend_hide(),
  size.chart = tmap::tm_chart_none(),
  size.free = NA,
  plot.order = tmap::tm_plot_order("size", reverse = FALSE),
  options = opt_tm_cartogram(),
  ...
)

tm_cartogram_ncont(
  size = 1,
  size.scale = tm_scale(),
  size.legend = tm_legend_hide(),
  size.chart = tm_chart_none(),
  size.free = NA,
  plot.order = tm_plot_order("size", reverse = FALSE),
  options = opt_tm_cartogram_ncont(),
  ...
)

tm_cartogram_dorling(
  size = 1,
  size.scale = tm_scale(),
  size.legend = tm_legend_hide(),
  size.chart = tm_chart_none(),
  size.free = NA,
  plot.order = tm_plot_order("size", reverse = FALSE),
  options = opt_tm_cartogram_dorling(),
  ...
)

opt_tm_cartogram(type = "cont", itermax = 15, ...)

opt_tm_cartogram_ncont(type = "ncont", expansion = 1, inplace = FALSE, ...)

opt_tm_cartogram_dorling(type = "dorling", share = 5, itermax = 1000, ...)
}
\arguments{
\item{size, size.scale, size.legend, size.chart, size.free}{Visual variable that specifies the polygon sizes.}

\item{plot.order}{Specification in which order the spatial features are drawn.
See [tmap:tm_plot_order()] for details.}

\item{options}{passed on to the corresponding `opt_<layer_function>` function}

\item{...}{arguments passed on to [cartogram::cartogram_cont()]}

\item{type}{cartogram type, one of: "cont" for contiguous cartogram, "ncont" for non-contiguous cartogram and "dorling" for Dorling cartograms}

\item{itermax, }{maximum number of iterations (see [cartogram::cartogram_cont()])}

\item{expansion}{factor expansion, see [cartogram::cartogram_ncont()] (argument `k`)}

\item{inplace}{should each polygon be modified in its original place? (`TRUE` by default)}

\item{share}{share of the bounding box filled with the larger circle (see [cartogram::cartogram_dorling()] argument `k`)}
}
\description{
Map layer that draws a cartogram. It is recommended to specify a proper crs in `tmap:tm_shape()`.
}
